Added the User Input Button
I have added the User Input Button that goes to the first time user input page.
This page is only the layout and it is a placeholder.
The information is not stored anywhere yet.
The Submit button will return to the main page.
The Return to tabs page (button from fragment to activity) is implemented different from a button in activity.

TODO:
Next and Previous buttons do not work. I have not attempted fixing this yet but it seems like an easy fix.
Still need to connect the user input data to the ROOM Database.