Functional Database and User Input
The User Input is now functional and works with the database
Some modifications were done to the UserDao.java
The PrimaryKey in the User was set to AutoGenerate = false because we are going to allow only one user at a time. If a new user is created, it will overwrite the old one.
homepage.xml was modified to show the user data input
HomePage.java was added to make the homepage.xml work
UserInput added the firstName, lastName, Age listeners and save them to the DataBase.

..Accidentalyl moved the FirstFragment.java, moved back to ui.main